+++ src/pyiem/plot/geoplot.py
@@ -1197,11 +1197,30 @@ class MapPlot:
         """Add overlay of filled state polygons"""
         geodf = load_geodf("us_states")
         if self.sector in ["nws", "conus"]:
-            st = ("NH", "MD", "CT")
-            geodf.loc[st, "lat"] = geodf.loc[st, "lat"] - 0.3
-            st = ("VT", "DE")
-            geodf.loc[st, "lat"] = geodf.loc[st, "lat"] + 0.3
+            st = ["NH", "MD", "CT"]
+            geodf.loc[st, "lat"] = geodf.loc[st, "lat"] - 0.3  # type: ignore
+            st = ["VT", "DE"]
+            geodf.loc[st, "lat"] = geodf.loc[st, "lat"] + 0.3  # type: ignore
         polygon_fill(self, geodf, data, **kwargs)
+        if self.sector != "nws":
+            return
+        # States that may have data, but are not plotted within sector=nws
+        msg = ["Not Plotted:"]
+        for st in "VI MH MP DC".split():  # eh, are there more?
+            if st in data:
+                val = kwargs.get("lblformat", "%s") % (data[st],)
+                msg.append(f"{st}: {val}")
+        if len(msg) > 1:
+            self.ax.text(
+                0.99,
+                0.4,
+                "\n".join(msg),
+                ha="right",
+                va="top",
+                fontsize=10,
+                bbox=dict(color="white"),
+                transform=self.ax.transAxes,
+            )
 
     def draw_fema_regions(self, color: str = "k", **kwargs):
         """Overlay FEMA Regions."""