+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this library are documented in this file.
 
 - Add `allowed_as_list` option to `iemapp()` helper to stop lists.
 - Add support for plotting by FEMA Regions.
+- Include simple table of un-plotted states for `MapPlot(sector="nws")` #967.
 - Introduce `radar_ptype` color ramp and `draw_radar_ptype_legend` for
 generating plots of HRRR ptype.